requirements.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2==2.9.1
requests==2.27.1
celery==5.2.3
python-dotenv==0.19.2
gunicorn==20.1.0
